third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-root-new-expected.txt

CONSOLE ERROR: line 19: Uncaught TypeError: div.createShadowRoot is not a function
This test checks the object 'host.createShadowRoot()' and internals.shadowRoot(div) are the same.

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
